requirements.txt
customtkinter>=5.2.0
requests>=2.28.0
paramiko>=3.0.0
Pillow>=10.0.0
cryptography>=41.0.0
darkdetect>=0.8.0
pyinstaller>=6.0.0
